nsupdate plugin: properly capture stderr
The nsupdate plugin calls the same-named binary using the
subprocess.Popen class, which by default, does NOT capture stderr
output.  Because of this, errors occurring during calls to nsupdate are
not properly reflected by ddupdate.

Set stderr=PIPE, and then check for any stderr output.  In the event of
any such output, print the error output, and return failure.

Signed-off-by: Antonio Russo <antonio.e.russo@gmail.com>