setup.py: New environment variable FINAL_PREFIX
Add new environment variable FINAL-PREFIX for use when packaging.
The variable replaces the root path used install.conf with the FINAL_PREFIX
prefix. Likewise, when using FINAL_PREFIX the otherwise absolute systemd
paths like /usr/lib/systemd becomes relative to the installation root.

The usecase is when packaging, which without this option has to
patch setup.py to cope with the installation paths.