setup.py: New environment variable FINAL_PREFIX
Add new environment variable  final-prefix, for use when packaging.
The variable replaces the otherwise used installation paths in
install.conf with paths with the final_prefix prefix. Likewise, when
using FINAL_PREFIX the otherwise absolute systemd paths like /usr/lib/systemd
becomes relative to the installation root.

The usecase is when packaging, which without this option has to
patch setup.py to cope with the installation paths.