Support multiple credentials for Google domains
Google domains dynamic dns API credentials are scoped to update a single
hostname.

To update multiple hosts, we update the way credentials are found in netrc.

To specify credentials that are specific to a target host, put the
following in ~/.netrc:

    machine mydomain.tld.domains.google.com.ddupdate login AAA password BBB
    machine mydomain2.tld.domains.google.com.ddupdate login CCC password DDD

With machine of the form:

    targethost.providerhost.ddupdate

Existing netrc entries like:

    domains.google.com login AAA password BBB

are still recognized but with lower priority than a per-host match.

Addresses #52 (only in the Google Domains case).