plugins: googledomains: Use configured authentication source.
After merging #23 the actual source for credentials could be
something else than the .netrc file. To cope with this, the library
function get_netrc_auth should be used despite that it has a bad
nam).

Re-factor the bundled get_netrc_auth to use the library function
instead of just the .netrc file.