New feature: run script on successful IP update
Added the ability to run an external script if the service plugin reports a successful IP update.
Only the cloudflare plugin has been updated to support this feature.
The script file can be specified in the config file, or passed as argument on the command line.
The script will be run with the following command line:
$ /path/to/script/file.sh "hostname" "New IPv4 or None" "New IPv6 or None"